# tests/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "B3 DateTime API",
    "description": "API para consultar horários de operação e dias de negociação da B3 (Bolsa de Valores de São Paulo).\n\n## Características\n\n* **Horários de Operação**: abertura e fechamento lidos do Redis, com cache local de 1 h\n* **Dias de Negociação**: dias úteis na B3 segundo o calendário BVMF do `exchange_calendars`\n* **Janela de dados**: o calendário cobre uma janela móvel; consulte `GET /v1/calendar-info`\n  para os limites vigentes\n* **Timezone**: todos os horários e datas usam America/Sao_Paulo\n\n## Códigos de resposta\n\n| Código | Quando |\n|---|---|\n| `200` | Sucesso |\n| `400` | Período inválido em `/v1/trading-days`: ordem, span acima do máximo ou fora da janela |\n| `404` | Redis disponível, mas a chave não existe |\n| `422` | Data mal formada |\n| `502` | Valor no Redis fora do formato `HH:MM` |\n| `503` | Redis indisponível sem cache válido, ou calendário indisponível |\n\nAs URLs canônicas não têm barra final (`/docs/` responde 404). Código, guia de migração e\nCHANGELOG: [github.com/rlquilez/b3datetime](https://github.com/rlquilez/b3datetime).\n\n## Autenticação\n\n**Sem autenticação no momento.** As requisições não precisam de nenhum header. Quando o Kong passar a exigir o header `apikey`, a API é publicada com `API_KEY_REQUIRED=true` e esta documentação passa a declarar o esquema de segurança.",
    "contact": {
      "name": "Rodrigo Quilez",
...
    },
    "license": {
...
    },
...
  },
  "paths": {
...
    "/v1/calendar-info": {
      "get": {
        "tags": [
          "Dias de Negociação"
        ],
        "summary": "Consultar os limites do calendário",
        "description": "Retorna os limites vigentes do calendário carregado.\n\nA janela do calendário é **móvel** e se desloca conforme o tempo passa, portanto\nconsulte este endpoint em vez de assumir uma data mínima fixa. Períodos fora destes\nlimites são rejeitados com 400 por `/v1/trading-days`.\n\n`coverage_*` é o intervalo respondível; `first_session`/`last_session` são o primeiro e\no último pregão dentro dele. Os dois diferem quando a janela começa num feriado ou fim\nde semana.",
        "operationId": "get_calendar_info_v1_calendar_info_get",
        "responses": {
          "200": {
            "description": "Limites obtidos com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarInfoResponse"
                },
                "example": {
                  "exchange": "BVMF",
                  "coverage_start": "2016-09-04",
                  "coverage_end": "2027-09-03",
                  "first_session": "2016-09-05",
                  "last_session": "2027-09-03",
                  "sessions_count": 2730,
                  "max_range_days": 3660
                }
              }
            }
          },
          "503": {
            "description": "Calendário de negociação indisponível",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "Service Unavailable",
                    "message": "Calendário de negociação indisponível"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/is-trading-day": {
      "get": {
        "tags": [
          "Dias de Negociação"
        ],
        "summary": "Verificar se hoje é dia de negociação",
        "description": "Verifica se o dia atual é um dia de negociação na B3, considerando feriados e\nfinais de semana do calendário BVMF.\n\nO \"dia atual\" é determinado no timezone America/Sao_Paulo. Se ele estiver fora da\njanela do calendário a resposta é 503 — nunca `false`, que seria indistinguível de um\nferiado legítimo.",
        "operationId": "is_trading_day_v1_is_trading_day_get",
        "responses": {
          "200": {
            "description": "Verificação realizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradingDayResponse"
                },
                "examples": {
                  "trading_day": {
                    "summary": "Dia de negociação",
                    "value": {
                      "date": "2026-09-04",
                      "is_trading_day": true
                    }
                  },
                  "non_trading_day": {
                    "summary": "Não é dia de negociação",
                    "value": {
                      "date": "2026-09-05",
                      "is_trading_day": false
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Calendário de negociação indisponível",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "Service Unavailable",
                    "message": "Calendário de negociação indisponível"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/trading-days": {
      "get": {
        "tags": [
          "Dias de Negociação"
        ],
        "summary": "Listar dias de negociação em um período",
        "description": "Retorna os dias de negociação (ou de não-negociação) num período.\n\n**Parâmetros**\n- `start`: data inicial no formato YYYY-MM-DD\n- `end`: data final no formato YYYY-MM-DD, maior ou igual a `start`\n- `exclude`: se `true`, retorna os dias **sem** negociação; se `false` (padrão), os\n  dias **com** negociação\n\n**Restrições**\n- O período deve estar inteiramente dentro da janela do calendário. Consulte\n  `GET /v1/calendar-info` para os limites vigentes. Fora dela a resposta é 400 — a API\n  não devolve resultado parcial em silêncio.\n- O span máximo por requisição é limitado (`max_range_days`); períodos maiores são\n  rejeitados com 400.\n- Data mal formada responde 422.",
        "operationId": "get_trading_days_v1_trading_days_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Data inicial (YYYY-MM-DD)",
              "title": "Start"
            },
            "description": "Data inicial (YYYY-MM-DD)",
            "examples": {
              "exemplo": {
                "summary": "Início do período",
                "value": "2026-09-01"
              }
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Data final (YYYY-MM-DD)",
              "title": "End"
            },
            "description": "Data final (YYYY-MM-DD)",
            "examples": {
              "exemplo": {
                "summary": "Fim do período",
                "value": "2026-09-07"
              }
            }
          },
          {
            "name": "exclude",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Se true, retorna dias SEM negociação; se false, dias COM negociação",
              "default": false,
              "title": "Exclude"
            },
            "description": "Se true, retorna dias SEM negociação; se false, dias COM negociação"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de datas obtida com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Trading Days V1 Trading Days Get"
                },
                "examples": {
                  "trading_days": {
                    "summary": "Dias de negociação",
                    "value": [
                      "2026-09-01",
                      "2026-09-02",
                      "2026-09-03",
                      "2026-09-04"
                    ]
                  },
                  "non_trading_days": {
                    "summary": "Dias sem negociação (exclude=true)",
                    "value": [
                      "2026-09-05",
                      "2026-09-06",
                      "2026-09-07"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Período inválido ou fora da janela do calendário",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "Bad Request",
                    "message": "Período fora da janela coberta pelo calendário (2016-09-04 a 2027-09-03). Consulte GET /v1/calendar-info para os limites vigentes."
                  }
                }
              }
            }
          },
          "503": {
            "description": "Calendário de negociação indisponível",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "Service Unavailable",
                    "message": "Calendário de negociação indisponível"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendarInfoResponse": {
        "properties": {
          "exchange": {
            "type": "string",
            "title": "Exchange",
            "description": "Código da bolsa no exchange_calendars",
            "example": "BVMF"
          },
          "coverage_start": {
            "type": "string",
            "title": "Coverage Start",
            "description": "Primeira data respondível. Períodos que comecem antes são rejeitados",
            "example": "2016-09-04"
          },
          "coverage_end": {
            "type": "string",
            "title": "Coverage End",
            "description": "Última data respondível",
            "example": "2027-09-03"
          },
          "first_session": {
            "type": "string",
            "title": "First Session",
            "description": "Primeiro dia de negociação dentro da cobertura",
            "example": "2016-09-05"
          },
          "last_session": {
            "type": "string",
            "title": "Last Session",
            "description": "Último dia de negociação dentro da cobertura",
            "example": "2027-09-03"
          },
          "sessions_count": {
            "type": "integer",
            "title": "Sessions Count",
            "description": "Quantidade de dias de negociação na cobertura",
            "example": 2730
          },
          "max_range_days": {
            "type": "integer",
            "title": "Max Range Days",
            "description": "Span máximo aceito por /v1/trading-days, em dias",
            "example": 3660
          }
        },
        "type": "object",
        "required": [
          "exchange",
          "coverage_start",
          "coverage_end",
          "first_session",
          "last_session",
          "sessions_count",
          "max_range_days"
        ],
        "title": "CalendarInfoResponse",
        "description": "Limites vigentes do calendário carregado."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TradingDayResponse": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date",
            "description": "Data verificada no formato YYYY-MM-DD",
            "example": "2026-09-04"
          },
          "is_trading_day": {
            "type": "boolean",
            "title": "Is Trading Day",
            "description": "Se é um dia de negociação na B3",
            "example": true
          }
        },
        "type": "object",
        "required": [
          "date",
          "is_trading_day"
        ],
        "title": "TradingDayResponse",
        "description": "Resultado da verificação de um dia de negociação."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
  "servers": [
...
}